aio-grab: use correct chroma buffer limit
We cannot assume that the number of chroma lines
equals half the number of luma lines.
There are guaranteed to be more luma lines than
the vertical resolution, and more chroma lines
than half the vertical resolution.
But that does not make their relation exactly 2:1.
Use the reported number of chroma lines instead.