Print messages to stderr instead of stdout
Output all that debug talk to stderr. This frees up stdout so you can
grab to a stream instead, which might be very useful.

Remove a few old debug lines. And don't mention the version number,
it has been 1.0 since the dawn of ages, so no one seems to care.